Fix dtype mismatch, wandb compat, and add smoke test for all methods
- PD: cast v_student and v_tgt to float32 before MSE loss
- wandb: replace finish_previous with explicit wandb.finish() (v0.25 compat)
- Add smoke_test_all.py: runs 10 steps of PD/CD/DMD1/DMD2, verifies finite loss

+++ distill_methods/src/progressive_distillation.py
@@ -212,6 +212,6 @@ class ProgressiveDistillation(BaseDistiller):
         # Previous implementation used x-space MSE, which scales as (2h)^2
         # and causes implicit loss weighting differences across stages.
         v_tgt = (x_tgt.detach() - x_t) / (2.0 * h)
-        loss = torch.nn.functional.mse_loss(v_student, v_tgt)
+        loss = torch.nn.functional.mse_loss(v_student.float(), v_tgt.float())
 
         return {"loss": loss}
